EchoSplit/requirements.txt
numpy==1.23.5
librosa<=0.9.2
matplotlib
soundfile
pydub
pygame
kivy
kivy_deps.angle
kivy_deps.glew
kivy_deps.sdl2
spleeter
eyed3
